requirements.txt
# requirements.txt
streamlit
sympy
google-generativeai
streamlit-option-menu